# ticket-shop-api/swagger.yaml
openapi: 3.0.0
info:
  title: API de Venda de Ingressos para Testes QA
  description: Uma API para gerenciamento de venda de ingressos para eventos
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor de Desenvolvimento
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Login:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            nome:
              type: string
            email:
              type: string
              format: email
            role:
              type: string
            dataCriacao:
              type: string
              format: date-time
        token:
          type: string
...
    Error:
      type: object
      properties:
        erro:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              param:
                type: string
              location:
                type: string
paths:
  /auth/register:
    post:
      summary: Registrar novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - email
                - senha
              properties:
                nome:
                  type: string
                email:
                  type: string
                  format: email
                senha:
                  type: string
                  format: password
                telefone:
                  type: string
                cpf:
                  type: string
      responses:
        '201':
          description: Usuário registrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: Realizar login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - senha
              properties:
                email:
                  type: string
                  format: email
                senha:
                  type: string
                  format: password
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Login'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...